refactor remove class and add test codes
