+++ src/main/java/com/employee/app/domain/vo/Grade.java
@@ -1,4 +1,4 @@
-package com.employee.app.domain;
+package com.employee.app.domain.vo;
 
 import java.util.Arrays;
