+++ src/main/java/com/employee/app/domain/vo/Name.java
@@ -1,4 +1,4 @@
-package com.employee.app.domain;
+package com.employee.app.domain.vo;
 
 import static com.employee.app.utils.StringUtils.*;
