+++ src/main/java/com/employee/app/view/ConsoleInput.java
@@ -2,10 +2,10 @@ package com.employee.app.view;
 
 import java.util.Scanner;
 
-import com.employee.app.domain.Email;
-import com.employee.app.domain.Grade;
-import com.employee.app.domain.Name;
-import com.employee.app.domain.Phone;
+import com.employee.app.domain.vo.Email;
+import com.employee.app.domain.vo.Grade;
+import com.employee.app.domain.vo.Name;
+import com.employee.app.domain.vo.Phone;
 
 public class ConsoleInput implements InputView {
 	private static final Scanner SCANNER = new Scanner(System.in);
