+++ src/test/java/com/employee/app/domain/vo/PhoneTest.java
@@ -1,4 +1,4 @@
-package com.employee.app.domain;
+package com.employee.app.domain.vo;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,13 +8,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import com.employee.app.utils.exception.InvalidInputException;
 
-class EmailTest {
+class PhoneTest {
 
-	@DisplayName("이메일이 아닌 값을 입력할 경우")
+	@DisplayName(value = "전화번호가 아닌 값을 입력할 경우")
 	@ParameterizedTest
-	@ValueSource(strings = {"aaa.com", "abc", "abc@@"})
-	void validateEmail(String email) {
+	@ValueSource(strings = {"11111", "010-333"})
+	void testValidatePhone (String phoneNumber) {
 		assertThatExceptionOfType(InvalidInputException.class)
-			.isThrownBy(() -> new Email(email));
+			.isThrownBy(() -> new Phone(phoneNumber));
 	}
 }
